+++ src/type_pool.h
@@ -208,7 +208,7 @@ public:
                                  std::unordered_map<std::string, HandlePtr> &&handles);
 
   TypeOrError createCLIRecordType(const std::string &typeName, ModId belongedModId,
-                                  CLIRecordType::Attr attr);
+                                  CLIRecordType::Attr attr, std::string &&desc);
 
   TypeOrError finalizeCLIRecordType(const CLIRecordType &recordType,
                                     std::unordered_map<std::string, HandlePtr> &&handles,